internal: Move internal testing code to internal/test
