+++ internal/test/eventlogs/arch-linux-workstation.bin
