+++ internal/test/eventlogs/rhel8-uefi.bin
