+++ internal/test/eventlogs/ubuntu-2104-no-secure-boot.bin
