+++ internal/test/test_other.go
@@ -1,6 +1,5 @@
 // +build !windows
-
-package internal
+package test
 
 import (
 	"flag"
