+++ internal/test/test_tpm.go
@@ -1,4 +1,4 @@
-package internal
+package test
 
 import (
 	"io"
